Should make this branch fast enough
Adds a global for controlling the value of a unit per building that unit could be used to construct for each city it could be used in.  This should speed things up and give us some workability here.  NOTE: if we make a unit capable of adding a 'bad' building or one that would replace a building and be worse than the one it's replacing, then this will become a misleading evaluation and more code logic in a time sensitive spot may become necessary, therefore caching would be advised once implemented.