Polished the plot list a bit.
Scroll buttons could appear when there was no reason to be able to scroll.
Sometimes there was an entirely empty row at the bottom of the plot list, like it thought there was supposed to be more units filling that row.
Was caused by a bit of counting confusion between units on the plot and units actually visible to the human player.