Espionage system cost adjustments and new globals to modify final amounts across the board.
Infiltration missions have been given a global to influence the final espionage they generate against the target player and are immediately being halved due to feedback on strength of how much they bring in.
A 10% increase in trade mission results.
A new global to influence the final cost of all espionage missions has been added and is increasing them all by 25% at the moment.  There are claims of some missions being much cheaper than they should be and I will need a save to see this as the xml and code don't support that conclusion so far as I can see on a pure read-through of the mechanism.