Feature: Makes the total adjacent damage visible in the city defense help tooltip
I'm sorry for forgetting the proper formatting for the commit message.  Really need to practice that with the proper information on-hand.  Anyhow, this adds a statement to the City Defense Tooltip (the Rook symbol in the top right of the city screen) that displays the current total amount of Adjacent Damage the city can put out.