Fix: Recalculation Process now considers whether Buildings are Obsoleted or Replaced when processing buildings.
Recalc checks buildings for obsoletion and will send those buildings through processing with the correct 'call' so as to capture the right tags to process only.  It also now checks for buildings being replaced by existing buildings in the city already and does not further process those.  This will fix a LOT of issues after a recalculation (Shift-CTRL-T) on any game in progress and will stop recalculations from creating very serious calculation errors.  The downside is significantly extended processing time in recalculation but that's not something one has to do all that often anyhow and does not slow down normal game processing itself at all.