fixed towers losing terrain control on upgrade
Discrepancy between setOwner and setImprovementType; culture was updating before new improvement was placed, so ownership fell through when improvement failed isActsAsCity check.
This has a possibility of causing oddities in edge cases when placing improvements that generate culture (cottages) in range of/replacing a fort, so keep an eye out?