CI: Run pylint in Github Actions
This commit adds `pylint` linting as a job for `Github Actions`, making
it easier to discover python errors while developing for this project.
The pylint scripts were added in commit:
`b0fcc7a97ad3a25610782fb9c1a0daf61ed3301f`

Pylint is run in two separate jobs as a matrix in GA, so it is easier to
see what version of python had an error. It also speeds up the CI.

Rejected designs:
Run pylint in `Appveyor`. This seems to complicate the appveyor yaml
file too much. The GA solution seems to be much more straightforward.

Other changes:
 * Ignored more `python 2` errors that was not detected during local
 runs
 * Ignored more `python 3` errors that was not detected during local
 runs (not added by this commits changes)
 * Changed the type of the `--python-version` argument otherwise it
 would fail when given a single version to lint
 * Replace the requirements.txt files used for installing `pylint`, it
 is unlikely that there ever will be more dependencies for this specific
 linting task