CI: Add unit test framework
Adds a framework for running unit tests written in `python 3`. Having
unit tests ensures that code does not break during refactoring, and
ensures that functions does what they advertise (reduces the chance of
bugs).

`pytest` was chosen for this task since it is arguably one of the most
widespread unit test framework for python.

By default pytest collections all files following this pattern:
`test_*.py`. So in order to add a test for some `python 3` code it is
as easy as creating a file `test_my_file.py` and add tests by creating
functions starting with `test_*` as function names. For example:

```python
def test_my_function():
    assert True
```

Since the test reports are not needed at this point, all of the output
of the tests are directly printed to `stdout`.

Also added a 400 seconds timeout per test (not per test file, but per
individual test), just in case a test waits for input/hangs.

This test setup is vulnerable to accidental test deletion. But even with
this drawback it is still better to have some tests, instead of none.

Another drawback is that this system is unlikely to work for
`python 2.4`, so as of now `python 2.4` remains without a test
framework.