Python 3: Fix broad-except error
Removed the entire try-except block since it is better to fail in such a
way that the user of the function `find_all_functions` knows that
something went wrong, instead of failing silently.

This fix is required for enabling `syntax-error` triggerd by `pylint`.
Otherwise, when the error is enabled the linter will complain about
this.