Fix: Quality Up Promotion requirements
Quality Up Promotions were being disqualified by units having more experience than the exact amount they needed.  Added a level offset to the XP calculator for the prerequisite check and cleaned up some style methods that were causing problems with breakpoint evaluations earlier in isPromotionValid.  (When we put a return line it must be on its OWN numbered line rather than coming in behind the closing bracket of the function preceding the return statement or you cannot put a clean stop on that return to check if that's the return being hit - it will ALWAYS stop there since it's evaluating the end bracket at that line.)