Quickfix to Categories Project Launch
One incorrect reference to the location of the Categories file that assumed it would have its own folder required fixing.  I did manage to get this tested in loading and running - not high risk as it is because it just loads and stores the data and maintains some new tags - nothing really in terms of application or display yet.  So if it CAN load without error, which it appears to be able to do, we should be good at the moment.