New Leaderhead Infos tags: DefaultTraits and DefaultComplexTraits
We can temporarily still use the original <Traits> tag to designate default traits on leaders but now we have the ability to use these tags to override those designations if we wish.  Eventually, we should use these tags as absolute replacements to the original <Traits> tag (with the Array and bool usage) but I wanted to give us time to adapt the xml first and investigate into how the python may still, and possibly quite erroneously for the sake of Developing Leaders, be using the original <Traits> as exposed to py.  We may need to expose these new vector tags to python, while perhaps debugging for Developing Leaders.  At least this gives us immediately the ability to add new Leader defaults for Complex Traits, which has a much wider set of options and not all are even exact smooth replacements as there are a few tags that are positive that are negative in Complex and so on.  If we didn't give this tag for our leaders, we'd have all kinds of gaps in the leader defaults when using Complex Traits and it's my hope that we can average out the default trait flavors on the options in use to automatically determine flavor settings on Leaders soon - making it easier for us with new leaders because we won't need to even define new leader flavors, just their default traits in core and complex arrangements.  This is part of how we can also move forward into totally random leader entities IN-GAME eventually for future projects like replaceable leaders and leaders as units.  Also lets us add leaders a LOT faster (or adapt a bunch of them I should say.)