Mormon Shrine -> Mormon Tabernacle
"Mormon Shrine" is just not a thing that exists. The Salt Lake Temple
might be a more obvious choice, but Discord seemed to feel it was
a bad name by virtue of being named after a specific geographical
feature (which, if added, might appear somewhere entirely different
on the map).

The Tabernacle was the first major building completed on Temple
Square and hence arguably can be their Holy City building on that
basis. The French and Italian texts are just the leads of the WP
articles edited to remove the mention of Salt Lake. Fortunately,
half the Polish text was already about the Tabernacle.

I am unable to either tackle the Russian text or the art. Sorry.
The existing art is pretty generic.