Human mod typos
Most of these were a curious aversion to the words "a" and "the".

For some of them it's (still) not very clear what they even mean,
alas.

"temperature of hundreds kelvins" was replaced in Heat Tolerance.
Almost everywhere on Earth save the coldest bits of Antarctica is
at a temperature of hundreds of Kelvins.